+++ src/entity/billDish.ts
@@ -30,6 +30,6 @@ export class BillDish extends BaseEntity {
     @Column('smallint', { default: 1 })
     public quantity: number;
 
-    @Column('money', { nullable: true })
+    @Column('float', { nullable: true })
     public price: number;
 }
