+++ src/entity/discountCode.ts
@@ -21,13 +21,13 @@ export class DiscountCode extends BaseEntity {
     @Column('timestamp with time zone')
     public endAt: Date;
 
-    @Column('money', { nullable: true })
+    @Column('float', { nullable: true })
     public minBillPrice: number;
 
     /**
      *   Maximum price this discount can reduce for a bill.
      */
-    @Column('money', { nullable: true })
+    @Column('float', { nullable: true })
     public maxPriceDiscount: number;
 
     @Column({ nullable: true })
